LayoutTests/inspector/elements-panel-structure.html

<html>
<head>
<script src="inspector-test.js"></script>
<script src="elements-tests.js"></script>
<script>

function doit()
{
    function callback(result)
    {
        var output = document.getElementById("outputPre");
        output.textContent = result.replace(/\u200b/g, "").replace(/\n/g, "").replace(/</g, "\n<");
        notifyDone();
    }
    evaluateInWebInspector("frontend_dumpAllElements", callback);
}


// Frontend functions.

function frontend_dumpAllElements(testController)
{
    testController.waitUntilDone();
    // Need test to be async to expand whole the tree first.
    frontend_expandDOMSubtree(WebInspector.domAgent.document);
    testController.runAfterPendingDispatches(function() {
        testController.notifyDone(WebInspector.panels.elements.treeOutline.element.textContent);
    });
}

</script>
</head>

<body onload="onload()">
<p>
Tests that elements panel shows dom tree structure.
</p>

<pre id="outputPre">
</pre>

</body>
</html>
