LayoutTests/inspector/elements-panel-structure-expected.txt

Tests that elements panel shows dom tree structure.


<html>
<head>
<body onload="onload()">
<p>Tests that elements panel shows dom tree structure.
</p>
<pre id="outputPre">
<div id="output">
</body>
</html>
